backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.12.5
pydantic-ai==1.44.0
python-dotenv==1.0.0
httpx==0.28.1
fpdf2==2.7.9
